created a cleaner version of redux
