+++ .gitignore
@@ -1,0 +1,6 @@
+node_modules/**/*
+
+# macOS
+.DS_Store
+yarn.lock
+.env
