+++ App.js
@@ -1,12 +1,13 @@
-import React, {  useState } from 'react'
-import {  StyleSheet } from 'react-native';
-import TodoList from './src';
+import React from "react";
+import { StyleSheet } from "react-native";
+import { Provider } from "react-redux";
+import { store } from "./src/redux/store";
+import TodoList from "./src/screens";
 
 export default function App() {
-  const [inputValues, setInputValues] = useState(['', '', '', '', '', '']);
-
-  return (
-      <TodoList />
-  )
+	return (
+		<Provider store={store}>
+			<TodoList />
+		</Provider>
+	);
 }
-const styles = StyleSheet.create({});
