--- src/actionTypes.tsx
@@ -1,3 +1,0 @@
-export  const ADD_TODO = 'addedNote';
-export  const REMOVE_TODO = 'removedNote';
-export const UPDATE_TODO = 'updateNote';
