+++ src/redux/actions/todoAction.ts
@@ -1,0 +1,19 @@
+import todoTypes from "../types/todoTypes";
+
+export const createTodo = (text: string) => {
+	return {
+		type: todoTypes.CREATE_TODO,
+		payload: {
+			text: text,
+		},
+	};
+};
+
+export const deleteTodo = (id: any) => {
+	return {
+		type: todoTypes.DELETE_TODO,
+		payload: {
+			id: id,
+		},
+	};
+};
