+++ src/redux/types/todoTypes.ts
@@ -1,0 +1,4 @@
+export default {
+	CREATE_TODO: "CREATE_TODO",
+	DELETE_TODO: "DELETE_TODO",
+};
