--- src/store.tsx
@@ -1,6 +1,0 @@
-import { createStore } from "redux";
-import reducer from "./reducer";
-
-const store = createStore(reducer);
-
-export default store;
